# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API Proyecto Final - Placey
  description: API de gestión de usuarios, alojamientos, reservas, comentarios y favoritos.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Servidor local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateUserRequest:
      type: object
      required: [name, email, password, phoneNumber, role, dateBirth]
      properties:
        name: { type: string, example: "carlos" }
        email: { type: string, example: "carlos@email.com" }
        password: { type: string, example: "123" }
        phoneNumber: { type: string, example: "12345" }
        role: { type: string, enum: [USER, ADMIN, HOST], example: "USER" }
        dateBirth: { type: string, format: date, example: "1900-10-10" }
    EditUserRequest:
      type: object
      properties:
        name: { type: string, example: "carlos" }
        phoneNumber: { type: string, example: "12345" }
        profilePhoto: { type: string, example: "url de la foto" }
    ChangePasswordRequest:
      type: object
      properties:
        password: { type: string, example: "Nueva123" }
...
paths:
  /users:
    post:
      tags: [Users]
      summary: Registro de usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        "201": { description: Usuario creado }
        "400": { description: Datos faltantes }
        "409": { description: Correo repetido }
  /users/{id}:
    get:
      tags: [Users]
      summary: Obtener usuario por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security: [ { bearerAuth: [] } ]
      responses:
        "200": { description: Datos del usuario }
        "401": { description: No autenticado }
        "403": { description: Sin permisos }
        "404": { description: No encontrado }
    put:
      tags: [Users]
      summary: Editar perfil del usuario
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditUserRequest'
      responses:
        "200": { description: Usuario actualizado }
        "400": { description: Datos faltantes }
        "401": { description: No autenticado }
        "403": { description: Sin permisos }
        "404": { description: No encontrado }
    delete:
      tags: [Users]
      summary: Eliminar usuario
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Usuario eliminado }
        "401": { description: No autenticado }
        "403": { description: Sin permisos }
        "404": { description: No encontrado }
    patch:
      tags: [Users]
      summary: Cambiar contraseña
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        "200": { description: Contraseña cambiada }
        "401": { description: No autenticado }
        "403": { description: Sin permisos }
        "404": { description: No encontrado }
...